requirements_mac.txt
numpy==1.24.3
tensorflow-macos==2.14.0
tensorflow-metal==1.1.0
pandas==2.2.3
scikit-learn==1.6.0
matplotlib==3.10.0
scipy==1.14.1
glob2==0.7
protobuf==4.25.5
seaborn==0.13.2
pyyaml==6.0.2
scikit-optimize==0.10.2